# swagger.json
{"openapi":"3.1.0","info":{"title":"SSAFY_HOME 3조","description":"SSAFY 부동산 프로젝트 API 문서","contact":{"name":"SSAFY","url":"https://github.com/ijip-naejip/backend"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Favorites","description":"관심 아파트 API"},{"name":"지역코드 API","description":"시도, 구군, 동 정보 조회 API"},{"name":"User API","description":"사용자 인증 및 회원 관리 API"},{"name":"헬스 체크","description":"서버 상태 확인 API"},{"name":"아파트 정보 API","description":"부동산 거래 내역 및 아파트 정보 조회 API"}],"paths":{"/user/update":{"put":{"tags":["User API"],"summary":"회원정보 수정","description":"로그인한 사용자의 정보를 수정합니다.","operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/signup":{"post":{"tags":["User API"],"summary":"회원가입","description":"새로운 사용자를 등록합니다.","operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/reset-password":{"post":{"tags":["User API"],"summary":"비밀번호 초기화 요청","description":"이메일로 임시 비밀번호를 전송합니다.","operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/reissue":{"post":{"tags":["User API"],"summary":"토큰 재발급","description":"Refresh Token을 이용하여 새로운 Access Token을 발급받습니다.","operationId":"reissue","parameters":[{"name":"refreshToken","in":"cookie","description":"리프레시 토큰 (쿠키)","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/logout":{"post":{"tags":["User API"],"summary":"로그아웃","description":"사용자를 로그아웃 처리합니다. 클라이언트 측에서 Access Token을 삭제해야 합니다.","operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/login":{"post":{"tags":["User API"],"summary":"로그인","description":"이메일과 비밀번호로 로그인하여 JWT 토큰을 발급받습니다.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/email-verification/request":{"post":{"tags":["User API"],"summary":"이메일 인증 코드 요청","description":"회원가입을 위한 이메일 인증 코드를 전송합니다.","operationId":"requestEmailVerification","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/email-verification/confirm":{"post":{"tags":["User API"],"summary":"이메일 인증 코드 확인","description":"전송된 이메일 인증 코드를 검증합니다.","operationId":"confirmEmailVerification","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/change-password":{"post":{"tags":["User API"],"summary":"비밀번호 변경","description":"로그인한 사용자의 비밀번호를 변경합니다.","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/api/v1/favorites":{"get":{"tags":["Favorites"],"summary":"관심 아파트 목록 조회","description":"로그인한 사용자가 등록한 관심 아파트 목록을 조회합니다.","operationId":"getFavorites","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseListFavoriteResponse"}}}}}},"post":{"tags":["Favorites"],"summary":"관심 아파트 추가","description":"새로운 관심 아파트를 등록합니다.","operationId":"addFavorite","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FavoriteAddRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseFavoriteResponse"}}}}}}},"/user/info":{"get":{"tags":["User API"],"summary":"회원정보 조회","description":"로그인한 사용자의 상세 정보를 조회합니다.","operationId":"getUserInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/api/v1/favorites/check":{"get":{"tags":["Favorites"],"summary":"관심 등록 여부 확인","description":"특정 아파트(평수 포함)가 관심 목록에 등록되어 있는지 확인합니다.","operationId":"checkFavorite","parameters":[{"name":"aptSeq","in":"query","required":true,"schema":{"type":"string"}},{"name":"pyung","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseBoolean"}}}}}}},"/api/v1/dongcode/{code}":{"get":{"tags":["지역코드 API"],"summary":"법정동 상세 정보 조회","description":"동코드를 이용하여 해당 법정동의 상세 정보(좌표, 주소 등)를 조회합니다","operationId":"getDetail","parameters":[{"name":"code","in":"path","description":"동코드 (예: 1168010100)","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseDongCodeResponse"}}}}}}},"/api/v1/dongcode/sido":{"get":{"tags":["지역코드 API"],"summary":"시도 목록 조회","description":"전국 모든 시도 목록을 조회합니다","operationId":"getSidoList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseListDongCodeResponse"}}}}}}},"/api/v1/dongcode/gugun":{"get":{"tags":["지역코드 API"],"summary":"구군 목록 조회","description":"선택한 시도에 해당하는 구군 목록을 조회합니다","operationId":"getGugunList","parameters":[{"name":"sido","in":"query","description":"시도명 (예: 서울특별시)","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseListDongCodeResponse"}}}}}}},"/api/v1/dongcode/dong":{"get":{"tags":["지역코드 API"],"summary":"동 목록 조회","description":"선택한 시도와 구군에 해당하는 동 목록을 조회합니다","operationId":"getDongList","parameters":[{"name":"sido","in":"query","description":"시도명 (예: 서울특별시)","required":true,"schema":{"type":"string"}},{"name":"gugun","in":"query","description":"구군명 (예: 강남구)","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseListDongCodeResponse"}}}}}}},"/api/v1/area":{"get":{"tags":["area-controller"],"summary":"사각형 범위 기반 지역 목록 조회","description":"사각형 좌표 범위에 따른 지역 목록을 조회합니다.\n- DONG/GUGUN/SIDO 범위: minPrice/maxPrice는 평당가(만원) 기준\n- APT/APT_DONG 범위: minPrice/maxPrice는 총 거래가(만원) 기준","operationId":"searchAreaAddress","parameters":[{"name":"minLat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"maxLat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"minLng","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"maxLng","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"scope","in":"query","required":true,"schema":{"type":"string","enum":["APT_DONG","APT","DONG","GUGUN","SIDO"]}},{"name":"minPrice","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"maxPrice","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"minPyung","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"maxPyung","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseListAddressResponse"}}}}}}},"/api/v1/apartments/{aptSeq}":{"get":{"tags":["아파트 정보 API"],"summary":"아파트별 상세 정보 조회","description":"특정 아파트의 상세 정보를 조회합니다","operationId":"getApartmentDetail","parameters":[{"name":"aptSeq","in":"path","description":"아파트 시퀀스 (예: 11680-1)","required":true,"schema":{"type":"string"}},{"name":"pyung","in":"query","description":"평형 (예: 22)","required":false,"schema":{"type":"string","default":"all"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseApartmentDetailResponse"}}}}}}},"/":{"get":{"tags":["헬스 체크"],"summary":"헬스 체크","description":"서버가 정상적으로 작동하고 있는지 확인합니다","operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseMapStringObject"}}}}}}},"/api/v1/favorites/{id}":{"delete":{"tags":["Favorites"],"summary":"관심 아파트 삭제 (ID)","description":"관심 아파트 ID(PK)를 이용하여 관심 목록에서 삭제합니다.","operationId":"removeFavorite","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseVoid"}}}}}}},"/api/v1/favorites/apt/{aptSeq}":{"delete":{"tags":["Favorites"],"summary":"관심 아파트 삭제 (aptSeq)","description":"아파트 고유번호(aptSeq)를 이용하여 관심 목록에서 삭제합니다.","operationId":"removeFavoriteByAptSeq","parameters":[{"name":"aptSeq","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponseVoid"}}}}}}}},"components":{"schemas":{"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"phone":{"type":"string"},"profileImage":{"type":"string"},"role":{"type":"string","enum":["ROLE_USER","ROLE_ADMIN"]},"socialType":{"type":"string","enum":["NONE","KAKAO","NAVER","GOOGLE"]},"socialId":{"type":"string"},"gender":{"type":"string","enum":["male","female","other"]},"ageGroup":{"type":"string","enum":["20s","30s","40s","50+"]},"job":{"type":"string","enum":["student","employee","business","freelancer","other"]},"maritalStatus":{"type":"string","enum":["single","married"]},"appPush":{"type":"boolean"},"emailNotification":{"type":"boolean"},"marketingNotification":{"type":"boolean"},"myHouseName":{"type":"string"},"myHouseAddress":{"type":"string"},"myHouseArea":{"type":"string"},"myHouseFloor":{"type":"string"},"myHousePrice":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"emailVerified":{"type":"boolean"}}},"FavoriteAddRequest":{"type":"object","description":"관심 아파트 추가 요청","properties":{"aptSeq":{"type":"string","description":"아파트 고유번호","example":"11680-102"},"aptName":{"type":"string","description":"아파트명","example":"래미안 강남"},"address":{"type":"string","description":"주소","example":"서울특별시 강남구 역삼동"},"pyung":{"type":"integer","format":"int32","description":"선택 평수","example":34},"dealAmount":{"type":"string","description":"거래가격","example":"15억 5000만원"},"latitude":{"type":"number","format":"double","description":"위도","example":37.5665},"longitude":{"type":"number","format":"double","description":"경도","example":126.978}}},"CommonResponseFavoriteResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/FavoriteResponse"}}},"FavoriteResponse":{"type":"object","description":"관심 아파트 정보","properties":{"id":{"type":"integer","format":"int64","description":"관심 목록 ID"},"aptSeq":{"type":"string","description":"아파트 고유번호"},"aptName":{"type":"string","description":"아파트명"},"address":{"type":"string","description":"주소"},"pyung":{"type":"integer","format":"int32","description":"평수"},"dealAmount":{"type":"string","description":"거래가격"},"latitude":{"type":"number","format":"double","description":"위도"},"longitude":{"type":"number","format":"double","description":"경도"},"buildYear":{"type":"integer","format":"int32","description":"건축년도"},"createdAt":{"type":"string","description":"등록일시"}}},"CommonResponseListFavoriteResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/FavoriteResponse"}}}},"CommonResponseBoolean":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"boolean"}}},"CommonResponseDongCodeResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/DongCodeResponse"}}},"DongCodeResponse":{"type":"object","properties":{"dongCode":{"type":"string"},"sidoName":{"type":"string"},"gugunName":{"type":"string"},"dongName":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"avgPrice":{"type":"integer","format":"int32"}}},"CommonResponseListDongCodeResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DongCodeResponse"}}}},"AddressResponse":{"type":"object","properties":{"dongCode":{"type":"string"},"sidoName":{"type":"string"},"gugunName":{"type":"string"},"dongName":{"type":"string"},"aptSeq":{"type":"string"},"aptName":{"type":"string"},"aptDong":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"avgPrice":{"type":"integer","format":"int32"},"primaryPyung":{"type":"integer","format":"int32"},"pricePerPyung":{"type":"integer","format":"int32"}}},"CommonResponseListAddressResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/AddressResponse"}}}},"ApartmentDetailResponse":{"type":"object","description":"아파트 상세 정보 데이터","properties":{"selectedPyung":{"type":"string","description":"선택된 평형 (all 또는 특정 평형 번호)","example":"all"},"apartmentInfo":{"$ref":"#/components/schemas/ApartmentInfoDto","description":"아파트 기본 정보"},"recentTransactions":{"type":"array","description":"최근 거래 내역 리스트","items":{"$ref":"#/components/schemas/RecentTransactionDto"}},"priceTrend":{"$ref":"#/components/schemas/PriceTrendDto","description":"6개월간 가격 변동 추이"}}},"ApartmentInfoDto":{"type":"object","description":"아파트 기본 정보","properties":{"aptSeq":{"type":"string","description":"아파트 고유번호","example":12345},"aptName":{"type":"string","description":"아파트 명","example":"잠실 엘스 아파트"},"address":{"type":"string","description":"주소","example":"서울특별시 송파구 올림픽로 99"},"avgPrice":{"type":"integer","format":"int32","description":"전체 평형 기준 평균 가격 (만 단위)","example":245000},"buildYear":{"type":"integer","format":"int32","description":"건축년도","example":2006},"latitude":{"type":"number","format":"double","description":"위도","example":37.5665},"longitude":{"type":"number","format":"double","description":"경도","example":126.978},"pyungTypes":{"type":"array","description":"아파트가 보유한 전체 평형 목록 (예: '22', '33')","example":["22","33"],"items":{"type":"string"}}}},"CommonResponseApartmentDetailResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ApartmentDetailResponse"}}},"PriceDataPointDto":{"type":"object","description":"월별 가격 데이터 포인트","properties":{"month":{"type":"string","description":"해당 월 (YYYY-MM)","example":"2025-05"},"avgPrice":{"type":"integer","format":"int32","description":"해당 월의 평균 가격 (만 단위)","example":245000},"transactionCount":{"type":"integer","format":"int32","description":"해당 월의 거래량","example":5}}},"PriceTrendDto":{"type":"object","description":"가격 변동 추이 데이터","properties":{"unit":{"type":"string","description":"기간 단위 (month만 가능)","example":"month"},"dataPoints":{"type":"array","description":"월별 가격 변동 데이터 포인트 리스트","items":{"$ref":"#/components/schemas/PriceDataPointDto"}}}},"RecentTransactionDto":{"type":"object","description":"최근 거래 내역 상세 (Record)","properties":{"transactionDate":{"type":"string","description":"거래 날짜 (YYYY-MM-DD)","example":"2025-05-15"},"pyungType":{"type":"string","description":"거래된 평형 타입","example":33},"dealAmount":{"type":"integer","format":"int32","description":"거래액 (만 단위)","example":245000},"floor":{"type":"integer","format":"int32","description":"층수","example":15},"aptDong":{"type":"string","description":"동 정보 (또는 동호수 정보)","example":101}}},"CommonResponseMapStringObject":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{}}}},"CommonResponseVoid":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}}}}}